requirements.txt
# GPS Image Localization Project - Dependencies
# Python 3.8+ required

# Core Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data Processing & Scientific Computing
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Image Processing
Pillow>=10.0.0
pillow-heif>=1.0.0
opencv-python>=4.8.0

# Visualization & Progress
matplotlib>=3.7.0
tqdm>=4.65.0

# Utilities
python-dotenv>=1.0.0

# Optional: For experiment tracking and visualization
# tensorboard>=2.13.0
# wandb>=0.15.0
# mlflow>=2.0.0
